Fix error with missing import for extension pod
The Kumulos.h header was included by the HTTP client code to make use of common error enums.

This file isn't included in the extension pod's file list so it was missing.

Move the errors to a shared errors header to resolve.
